feat: add markdown to all descriptions
