+++ packages/ui/src/Documentation/components/MemberDescription.tsx
@@ -1,4 +1,5 @@
 import { DocumentTextIcon } from "@heroicons/react/24/outline";
+import { ReadmeMarkdown } from "@website/markdown";
 import React from "react";
 import type { CommentParser } from "typedoc-json-parser";
 
@@ -12,11 +13,15 @@ export const MemberDescription: React.FC<Props> = ({ comment }) => {
 			<h2 className="flex items-center gap-2 text-6 font-medium">
 				<DocumentTextIcon className="h-6" /> Description
 			</h2>
-			<p>{comment.description}</p>
+			<div className="[&>div.markdown]:!font-mono [&>div.markdown]:px-0">
+				<ReadmeMarkdown content={comment.description} fullName="" version="" />
+			</div>
 			{comment.blockTags.map((tag) => (
 				<div key={tag.name} className="mt-2">
 					<h3 className="font-semibold text-5">{tag.name}</h3>
-					<p>{tag.text}</p>
+					<div className="[&>div.markdown]:!font-mono [&>div.markdown]:px-0">
+						<ReadmeMarkdown content={tag.text} fullName="" version="" />
+					</div>
 				</div>
 			))}
 		</section>
