+++ packages/ui/src/Documentation/components/index.ts
@@ -2,3 +2,4 @@ export * from "./MemberDescription";
 export * from "./MemberTitle";
 export * from "./MemberParameters";
 export * from "./MemberTypeParameters";
+export * from "./MemberProperties";
